[WINESYNC] mshtml: Fix HTMLWindow3_setInterval behavior on 0 msec intervals.

wine commit id 4e77355f3a94e00a826fd91c564c1a524d0b7093 by Piotr Caban <piotr@codeweavers.com>

+++ dll/win32/mshtml/mshtml_private.h
@@ -1144,7 +1144,7 @@ HRESULT push_task(task_t*,task_proc_t,task_proc_t,LONG) DECLSPEC_HIDDEN;
 void remove_target_tasks(LONG) DECLSPEC_HIDDEN;
 void flush_pending_tasks(LONG) DECLSPEC_HIDDEN;
 
-HRESULT set_task_timer(HTMLInnerWindow*,DWORD,BOOL,IDispatch*,LONG*) DECLSPEC_HIDDEN;
+HRESULT set_task_timer(HTMLInnerWindow*,LONG,BOOL,IDispatch*,LONG*) DECLSPEC_HIDDEN;
 HRESULT clear_task_timer(HTMLInnerWindow*,BOOL,DWORD) DECLSPEC_HIDDEN;
 
 const char *debugstr_mshtml_guid(const GUID*) DECLSPEC_HIDDEN;
